Ajuste produto pedido dto.

+++ src/main/java/com/ericson/controle_de_produtos_pedidos/dto/ProdutoPedidoDto.java
@@ -1,5 +1,7 @@
 package com.ericson.controle_de_produtos_pedidos.dto;
 
-public record ProdutoPedidoDto (Integer quantidade, Double precoProduto) {
+import java.math.BigDecimal;
+
+public record ProdutoPedidoDto (Integer quantidade, BigDecimal precoProduto) {
 
 }
